Add files via upload
A program in C that checks the validity of a given credit card number via Luhn's Algorithm, and then reports (via printf) whether it is a valid American Express, MasterCard, or Visa card number, per the definitions of each’s format herein. For given numbers that do not satisfy the specified definitions for AMEX, MASTERCARD or VISA, INVALID will be returned. 

Signed-off-by: Gracia Chew <159509493+gracias022@users.noreply.github.com>